chore: refactor code and tests
move token_factory into its own space
